setup: only setup nvm if not exist

+++ setup.sh
@@ -18,7 +18,9 @@ if [ "$1" == "as" ] ; then
   # Nodejs
   NVM_DIR="$BASEPATH/nvm"
   mkdir -p $NVM_DIR
-  curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.37.2/install.sh | bash
+  if [ ! -f $NVM_DIR/nvm.sh ]; then
+    curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.37.2/install.sh | bash
+  fi
   [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
   [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
   nvm install node
